Changes made.Digital,Analog & Delay done.

+++ headers/analog.h
@@ -1,2 +1,2 @@
 int cmd_analog_out(int h,int pin_no,double val);
-double cmd_analog_in(int h,int pin_no);
+unsigned int cmd_analog_in(int h,int pin_no);
